+++ renderer.h
@@ -28,5 +28,5 @@ class C_Renderer
 		void M_Draw();
 		void M_Use(const C_Shader& s);
 		void M_AddEntity(C_GfxEntity* e);
-		C_GfxEntity* M_GetEntity(unsigned int id);
+		C_GfxEntity* M_GetEntity(unsigned short id);
 };
